requirements.txt
python-telegram-bot==21.4
python-dotenv==1.0.1
httpx==0.27.2
